requirements.txt
opencv-python
pycurl
pillow
